requirements.txt
streamlit
streamlit-ace
resvg-py
python-pptx
openai
python-dotenv
pyyaml
svgelements
